requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.6
idna==3.4
mysqlclient==2.2.0
Pillow==10.0.1
py3rijndael==0.3.3
pycryptodome==3.19.0
PyJWT==2.8.0
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.6